wait for push to finish before trying another push
