+++ jni/hw/dji_display.c
@@ -1,10 +1,13 @@
 #include <stdlib.h>
 #include "dji_display.h"
+#include "util/debug.h"
 
 #define GOGGLES_V1_VOFFSET 575
 #define GOGGLES_V2_VOFFSET 215
 
 static duss_result_t pop_func(duss_disp_instance_handle_t *disp_handle,duss_disp_plane_id_t plane_id, duss_frame_buffer_t *frame_buffer,void *user_ctx) {
+    dji_display_state_t *display_state = (dji_display_state_t *)user_ctx;
+    display_state->frame_waiting = 0;
     return 0;
 }
 
@@ -15,6 +18,7 @@ dji_display_state_t *dji_display_state_alloc(uint8_t is_v2_goggles) {
     display_state->fb_1 = (duss_frame_buffer_t *)calloc(1,sizeof(duss_frame_buffer_t));
     display_state->pb_0 = (duss_disp_plane_blending_t *)calloc(1, sizeof(duss_disp_plane_blending_t));
     display_state->is_v2_goggles = is_v2_goggles;
+    display_state->frame_waiting = 0;
     return display_state;
 }
 
@@ -27,7 +31,6 @@ void dji_display_state_free(dji_display_state_t *display_state) {
 }
 
 void dji_display_close_framebuffer(dji_display_state_t *display_state) {
- 
     duss_hal_display_port_enable(display_state->disp_instance_handle, 3, 0);
     duss_hal_display_release_plane(display_state->disp_instance_handle, display_state->plane_id);
     duss_hal_display_close(display_state->disp_handle, &display_state->disp_instance_handle);
@@ -94,7 +97,7 @@ void dji_display_open_framebuffer(dji_display_state_t *display_state, duss_disp_
         printf("failed to acquire plane");
         exit(0);
     }
-    res = duss_hal_display_register_frame_cycle_callback(display_state->disp_instance_handle, plane_id, &pop_func, 0);
+    res = duss_hal_display_register_frame_cycle_callback(display_state->disp_instance_handle, plane_id, &pop_func, display_state);
     if (res != 0) {
         printf("failed to register callback");
         exit(0);
@@ -171,10 +174,117 @@ void dji_display_open_framebuffer(dji_display_state_t *display_state, duss_disp_
     }
 }
 
+
+void dji_display_open_framebuffer_injected(dji_display_state_t *display_state, duss_disp_instance_handle_t *disp, duss_hal_obj_handle_t ion_handle, duss_disp_plane_id_t plane_id) {
+    uint32_t hal_device_open_unk = 0;
+    duss_result_t res = 0;
+    display_state->disp_instance_handle = disp;
+    display_state->ion_handle = ion_handle;
+    display_state->plane_id = plane_id;
+
+    // PLANE BLENDING
+
+    display_state->pb_0->is_enable = 1;
+
+    // TODO just check hwid to figure this out. Not actually V1/V2 related but an HW version ID.
+
+    display_state->pb_0->voffset = GOGGLES_V1_VOFFSET;
+    display_state->pb_0->hoffset = 0;
+
+    // On Goggles V1, the UI and video are in Z-Order 1. On Goggles V2, they're in Z-Order 4.
+    // Unfortunately, this means we cannot draw below the DJI UI on Goggles V1. But, on Goggles V2 we get what we want.
+
+    display_state->pb_0->order = 2;
+
+    // Global alpha - disable as we want per pixel alpha.
+
+    display_state->pb_0->glb_alpha_en = 0;
+    display_state->pb_0->glb_alpha_val = 0;
+
+    // These aren't documented. Blending algorithm 0 is employed for menus and 1 for screensaver.
+
+    display_state->pb_0->blending_alg = 1;
+
+    // No idea what this "plane mode" actually does but it's different on V2
+    uint8_t acquire_plane_mode = display_state->is_v2_goggles ? 6 : 0;
+
+    DEBUG_PRINT("acquire plane\n");
+    res = duss_hal_display_aquire_plane(display_state->disp_instance_handle,acquire_plane_mode,&plane_id);
+    if (res != 0) {
+        DEBUG_PRINT("failed to acquire plane");
+        exit(0);
+    }
+    res = duss_hal_display_register_frame_cycle_callback(display_state->disp_instance_handle, plane_id, &pop_func, 0);
+    if (res != 0) {
+        DEBUG_PRINT("failed to register callback");
+        exit(0);
+    }
+
+    res = duss_hal_display_plane_blending_set(display_state->disp_instance_handle, plane_id, display_state->pb_0);
+
+    if (res != 0) {
+        DEBUG_PRINT("failed to set blending");
+        exit(0);
+    }
+    DEBUG_PRINT("alloc ion buf\n");
+    res = duss_hal_mem_alloc(display_state->ion_handle,&display_state->ion_buf_0,0x473100,0x400,0,0x17);
+    if (res != 0) {
+        DEBUG_PRINT("failed to allocate VRAM");
+        exit(0);
+    }
+    res = duss_hal_mem_map(display_state->ion_buf_0, &display_state->fb0_virtual_addr);
+    if (res != 0) {
+        DEBUG_PRINT("failed to map VRAM");
+        exit(0);
+    }
+    res = duss_hal_mem_get_phys_addr(display_state->ion_buf_0, &display_state->fb0_physical_addr);
+    if (res != 0) {
+        DEBUG_PRINT("failed to get FB0 phys addr");
+        exit(0);
+    }
+    DEBUG_PRINT("first buffer VRAM mapped virtual memory is at %p : %p\n", display_state->fb0_virtual_addr, display_state->fb0_physical_addr);
+
+    res = duss_hal_mem_alloc(display_state->ion_handle,&display_state->ion_buf_1,0x473100,0x400,0,0x17);
+    if (res != 0) {
+        DEBUG_PRINT("failed to allocate FB1 VRAM");
+        exit(0);
+    }
+    res = duss_hal_mem_map(display_state->ion_buf_1,&display_state->fb1_virtual_addr);
+    if (res != 0) {
+        DEBUG_PRINT("failed to map FB1 VRAM");
+        exit(0);
+    }
+    res = duss_hal_mem_get_phys_addr(display_state->ion_buf_1, &display_state->fb1_physical_addr);
+    if (res != 0) {
+        DEBUG_PRINT("failed to get FB1 phys addr");
+        exit(0);
+    }
+    DEBUG_PRINT("second buffer VRAM mapped virtual memory is at %p : %p\n", display_state->fb1_virtual_addr, display_state->fb1_physical_addr);
+
+    for(int i = 0; i < 2; i++) {
+        duss_frame_buffer_t *fb = i ? display_state->fb_1 : display_state->fb_0;
+        fb->buffer = i ? display_state->ion_buf_1 : display_state->ion_buf_0;
+        fb->pixel_format = display_state->is_v2_goggles ? DUSS_PIXFMT_RGBA8888_GOGGLES_V2 : DUSS_PIXFMT_RGBA8888; // 20012 instead on V2
+        fb->frame_id = i;
+        fb->planes[0].bytes_per_line = 0x1680;
+        fb->planes[0].offset = 0;
+        fb->planes[0].plane_height = 810;
+        fb->planes[0].bytes_written = 0x473100;
+        fb->width = 1440;
+        fb->height = 810;
+        fb->plane_count = 1;
+    }
+}
+
 void dji_display_push_frame(dji_display_state_t *display_state, uint8_t which_fb) {
     duss_frame_buffer_t *fb = which_fb ? display_state->fb_1 : display_state->fb_0;
     duss_hal_mem_sync(fb->buffer, 1);
-    duss_hal_display_push_frame(display_state->disp_instance_handle, display_state->plane_id, fb);
+    if (display_state->frame_waiting == 0) {
+        display_state->frame_waiting = 1;
+        duss_hal_display_push_frame(display_state->disp_instance_handle, display_state->plane_id, fb);
+    } else {
+        DEBUG_PRINT("!!! Dropped frame due to pending frame push!\n");
+    }
 }
 
 void *dji_display_get_fb_address(dji_display_state_t *display_state, uint8_t which_fb) {
