+++ jni/hw/dji_display.h
@@ -1,5 +1,4 @@
-#ifndef DJI_DISPLAY_H
-#define DJI_DISPLAY_H
+#pragma once
 #include "duml_hal.h"
 
 typedef struct dji_display_state_s {
@@ -18,12 +17,13 @@ typedef struct dji_display_state_s {
     duss_frame_buffer_t *fb_1;
     duss_disp_plane_blending_t *pb_0;
     uint8_t is_v2_goggles;
+    uint8_t frame_waiting;
 } dji_display_state_t;
 
 void dji_display_push_frame(dji_display_state_t *display_state, uint8_t which_fb);
 void dji_display_open_framebuffer(dji_display_state_t *display_state, duss_disp_plane_id_t plane_id);
+void dji_display_open_framebuffer_injected(dji_display_state_t *display_state, duss_disp_instance_handle_t *disp, duss_hal_obj_handle_t ion_handle, duss_disp_plane_id_t plane_id);
 void dji_display_close_framebuffer(dji_display_state_t *display_state);
 dji_display_state_t *dji_display_state_alloc(uint8_t is_v2_goggles);
 void dji_display_state_free(dji_display_state_t *display_state);
 void *dji_display_get_fb_address(dji_display_state_t *display_state, uint8_t which_fb);
-#endif
