+++ jni/net/network.h
@@ -1,2 +1,4 @@
+#pragma once
+
 int connect_to_server(char *address, int port);
 int bind_socket(int port);
