+++ jni/util/fs_util.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <stdint.h>
 
 int32_t get_int_from_fs(char* path);
