+++ jni/util/time_util.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <time.h>
 #include <stdbool.h>
